refactor(reducers): move `results`
