+++ app/renderer/src/reducers/results/index.js
@@ -1,0 +1,1 @@
+export { default } from './results';
